+++ modules/crowdfundings/lib/payments/stripe/webhookHandler.js
@@ -1,9 +1,6 @@
 const debug = require('debug')('crowdfundings:webhooks:stripe')
 const getStripeClients = require('./clients')
-const { sendMailTemplate } = require('@orbiting/backend-modules-mail')
-const _ = {
-  get: require('lodash/get')
-}
+const handlers = require('./webhooks/index')
 
 module.exports = async ({ pgdb, t }) => {
   const {
@@ -11,14 +8,10 @@ module.exports = async ({ pgdb, t }) => {
     connectedAccounts
   } = await getStripeClients(pgdb)
 
-  const typesOfIntereset = [
-    'invoice.payment_succeeded',
-    'invoice.payment_failed',
-    'charge.succeeded',
-    'charge.refunded',
-    'customer.subscription.deleted',
-    'customer.subscription.updated'
-  ]
+  const typesOfInterest = handlers.reduce(
+    (arr, handler) => [...arr, ...handler.eventTypes],
+    []
+  )
 
   return async ({
     req,
@@ -42,334 +35,11 @@ module.exports = async ({ pgdb, t }) => {
       return 400
     }
 
-    if (typesOfIntereset.indexOf(event.type) > -1) {
+    if (typesOfInterest.includes(event.type)) {
       debug('%O', event)
 
-      /*
-      const account = event.accountId
-        ? accounts.find( a => a.accountId === event.account )
-        : platform
-      if (!account) {
-        throw new Error("stripe handleWebhook didn't find local account for event")
-      }
-      */
-
-      // invoice.payment_succeeded includes:
-      // pledgeId, charge total and charge id
-      // but not the charge details, charge may not
-      // exist at the time this hook is received
-      if (event.type === 'invoice.payment_succeeded') {
-        const invoice = _.get(event, 'data.object')
-        const subscription = _.get(event, 'data.object.lines.data[0]')
-        if (subscription.type === 'subscription') {
-          const {
-            charge: chargeId,
-            total
-          } = invoice
-          const pledgeId = subscription.metadata.pledgeId
-
-          const transaction = await pgdb.transactionBegin()
-          try {
-            // synchronize with payPledge
-            const pledge = await transaction.query(`
-              SELECT *
-              FROM pledges
-              WHERE id = :pledgeId
-              FOR UPDATE
-            `, {
-              pledgeId
-            })
-              .then(response => response[0])
-              .catch(e => {
-                console.error(e)
-                return null
-              })
-
-            if (!pledge) {
-              console.warn(`received webhook for unknown pledgeId ${pledgeId}`)
-              await transaction.transactionRollback()
-              return 500
-            }
-
-            const existingPayments = await transaction.public.payments.find({
-              method: 'STRIPE',
-              pspId: chargeId
-            })
-
-            if (!existingPayments.length) {
-              // the first membershipPeriod was already inserted by generateMemberships
-              // but not the corresponding payment, save that here
-              const payment = await transaction.public.payments.insertAndGet({
-                type: 'PLEDGE',
-                method: 'STRIPE',
-                total: total,
-                status: 'PAID',
-                pspId: chargeId
-              })
-
-              await transaction.public.pledgePayments.insert({
-                pledgeId,
-                paymentId: payment.id,
-                paymentType: 'PLEDGE'
-              })
-            }
-
-            const memberships = await transaction.public.memberships.find({
-              pledgeId
-            })
-            const firstNotification = memberships[0].subscriptionId === null
-            const beginDate = new Date(subscription.period.start * 1000)
-            const endDate = new Date(subscription.period.end * 1000)
-            if (firstNotification) {
-              // remember subscriptionId
-              await transaction.public.memberships.update({
-                id: memberships.map(m => m.id)
-              }, {
-                subscriptionId: subscription.id
-              })
-              // synchronize beginDate and endDate with stripe
-              await transaction.query(`
-                UPDATE "membershipPeriods" mp
-                SET
-                  "webhookEventId" = :webhookEventId,
-                  "beginDate" = :beginDate,
-                  "endDate" = :endDate,
-                  "updatedAt" = :now
-                WHERE
-                  ARRAY[mp."membershipId"] && :membershipIds AND
-                  mp."webhookEventId" is null
-              `, {
-                webhookEventId: event.id,
-                beginDate,
-                endDate,
-                now: new Date(),
-                membershipIds: memberships.map(m => m.id)
-              })
-            } else {
-              // check for duplicate event
-              if (!(await transaction.public.membershipPeriods.findFirst({ webhookEventId: event.id }))) {
-                // insert membershipPeriods
-                await Promise.all(memberships.map(membership => {
-                  return transaction.public.membershipPeriods.insert({
-                    membershipId: membership.id,
-                    beginDate,
-                    endDate,
-                    webhookEventId: event.id
-                  })
-                }))
-                await transaction.public.memberships.update({
-                  id: memberships.map(m => m.id)
-                }, {
-                  active: true,
-                  updatedAt: new Date()
-                })
-              }
-            }
-            await transaction.transactionCommit()
-          } catch (e) {
-            await transaction.transactionRollback()
-            console.info('transaction rollback', { error: e })
-            console.error(e)
-            throw e
-          }
-        }
-      // stripe tried to charge the card but this failed
-      } else if (event.type === 'invoice.payment_failed') {
-        const subscription = _.get(event, 'data.object.lines.data[0]')
-        if (subscription.type === 'subscription') {
-          const pledgeId = subscription.metadata.pledgeId
-          const user = await pgdb.query(`
-            SELECT u.*
-            FROM users u
-            JOIN pledges p
-            ON u.id = p."userId" AND
-            p.id = :pledgeId
-          `, {
-            pledgeId
-          })
-            .then(response => response[0])
-            .catch(e => {
-              console.error(e)
-              return null
-            })
-
-          if (!user) {
-            throw new Error('user for invoice.payment_failed event not found! subscriptionId:' + subscription.id)
-          }
-
-          await sendMailTemplate({
-            to: user.email,
-            subject: t('api/email/subscription/payment/failed/subject'),
-            templateName: 'subscription_failed',
-            globalMergeVars: [
-              { name: 'NAME',
-                content: [user.firstName, user.lastName]
-                  .filter(Boolean)
-                  .join(' ')
-              }
-            ]
-          })
-        }
-      // charge.succeeded contains all the charge details
-      // but not the pledgeId
-      // if this event arrives before invoice.payment_succeeded
-      // we reject it and wait for the webhook to fire again
-      } else if (event.type === 'charge.succeeded') {
-        const charge = _.get(event, 'data.object')
-        const transaction = await pgdb.transactionBegin()
-        try {
-          const existingPayment = await transaction.query(`
-            SELECT *
-            FROM payments
-            WHERE "pspId" = :pspId
-            FOR UPDATE
-          `, {
-            pspId: charge.id
-          })
-            .then(response => response[0])
-            .catch(e => {
-              console.error(e)
-              return null
-            })
-
-          if (existingPayment) {
-            await transaction.public.payments.update({
-              id: existingPayment.id
-            }, {
-              pspPayload: charge,
-              updatedAt: new Date()
-            })
-          } else {
-            debug('no existing payment found in charge.succeeded. rejecting event %O', event)
-            await transaction.transactionRollback()
-            return 503
-          }
-
-          await transaction.transactionCommit()
-        } catch (e) {
-          await transaction.transactionRollback()
-          console.info('transaction rollback', { error: e })
-          console.error(e)
-          throw e
-        }
-      } else if (event.type === 'charge.refunded') {
-        const charge = _.get(event, 'data.object')
-        const transaction = await pgdb.transactionBegin()
-        try {
-          const existingPayment = await transaction.query(`
-            SELECT *
-            FROM payments
-            WHERE "pspId" = :pspId
-            FOR UPDATE
-          `, {
-            pspId: charge.id
-          })
-            .then(response => response[0])
-            .catch(e => {
-              console.error(e)
-              return null
-            })
-
-          if (existingPayment) {
-            await transaction.public.payments.update({
-              id: existingPayment.id
-            }, {
-              status: 'REFUNDED',
-              updatedAt: new Date()
-            })
-          } else {
-            debug('no existing payment found in charge.refunded. rejecting event %O', event)
-            await transaction.transactionRollback()
-            return 503
-          }
-
-          await transaction.transactionCommit()
-        } catch (e) {
-          await transaction.transactionRollback()
-          console.info('transaction rollback', { error: e })
-          console.error(e)
-          throw e
-        }
-      } else if (event.type === 'customer.subscription.updated' ||
-      event.type === 'customer.subscription.deleted') {
-        const subscription = event.data.object
-        const pledgeId = subscription.metadata.pledgeId
-
-        const transaction = await pgdb.transactionBegin()
-        try {
-          const pledge = await transaction.query(`
-            SELECT *
-            FROM pledges
-            WHERE id = :pledgeId
-            FOR UPDATE
-          `, {
-            pledgeId
-          })
-            .then(response => response[0])
-            .catch(e => {
-              console.error(e)
-              return null
-            })
-
-          if (!pledge) {
-            throw new Error('pledge for customer.subscription event not found! subscriptionId:' + subscription.id)
-          }
-
-          if (subscription.cancel_at_period_end) {
-            await transaction.public.memberships.update({
-              pledgeId
-            }, {
-              renew: false,
-              updatedAt: new Date()
-            })
-          }
-
-          // Possible values are trialing, active, past_due, canceled, or unpaid
-          // https://stripe.com/docs/api/node#subscription_object
-          if (subscription.status === 'canceled') {
-            const existingMembership = await transaction.public.memberships.findFirst({
-              pledgeId
-            })
-            // membership might have been moved by cancelPledge
-            if (!existingMembership) {
-              await transaction.public.memberships.update({
-                pledgeId
-              }, {
-                active: false,
-                renew: false,
-                updatedAt: new Date()
-              })
-
-              const user = await transaction.public.users.findOne({
-                id: pledge.userId
-              })
-
-              if (existingMembership.active) {
-                await sendMailTemplate({
-                  to: user.email,
-                  subject: t('api/email/subscription/deactivated/subject'),
-                  templateName: 'subscription_end',
-                  globalMergeVars: [
-                    { name: 'NAME',
-                      content: [user.firstName, user.lastName]
-                        .filter(Boolean)
-                        .join(' ')
-                    }
-                  ]
-                })
-              }
-            }
-          }
-          await transaction.transactionCommit()
-        } catch (e) {
-          await transaction.transactionRollback()
-          console.info('transaction rollback', { error: e })
-          console.error(e)
-          throw e
-        }
-      } else {
-        throw new Error('missing handler for event type: ' + event.type)
-      }
+      const { handle } = typesOfInterest.find(handler => handler.eventTypes.includes(event.type))
+      return handle(event, pgdb)
     } else {
       debug(`webhookHandler ignoring event with type: ${event.type}`)
     }
